+++ plugins/woocommerce/includes/rest-api/Controllers/Version3/class-wc-rest-orders-controller.php
@@ -158,6 +158,8 @@ class WC_REST_Orders_Controller extends WC_REST_Orders_V2_Controller {
 		 * The dynamic portion of the hook name, `$this->post_type`,
 		 * refers to the object type slug.
 		 *
+		 * @since ?
+		 *
 		 * @param WC_Data         $order    Object object.
 		 * @param WP_REST_Request $request  Request object.
 		 * @param bool            $creating If is creating a new object.
@@ -244,7 +246,7 @@ class WC_REST_Orders_Controller extends WC_REST_Orders_V2_Controller {
 		$cpt_hidden_keys = array();
 
 		if ( OrderUtil::custom_orders_table_usage_is_enabled() ) {
-			$cpt_hidden_keys = (new \WC_Order_Data_Store_CPT())->get_internal_meta_keys();
+			$cpt_hidden_keys = ( new \WC_Order_Data_Store_CPT() )->get_internal_meta_keys();
 		}
 
 		// XXX: This might be removed once we finalize the design for internal keys vs meta vs props in COT.
@@ -308,35 +310,6 @@ class WC_REST_Orders_Controller extends WC_REST_Orders_V2_Controller {
 		}
 
 		return $args;
-	}
-
-	/**
-	 * Get objects.
-	 *
-	 * @param  array $query_args Query args.
-	 * @return array
-	 */
-	protected function get_objects( $query_args ) {
-		// Do not use WC_Order_Query for the CPT datastore.
-		if ( ! OrderUtil::custom_orders_table_usage_is_enabled() ) {
-			return parent::get_objects( $query_args );
-		}
-
-		$query   = new \WC_Order_Query(
-			array_merge(
-				$query_args,
-				array(
-					'paginate' => true,
-				)
-			)
-		);
-		$results = $query->get_orders();
-
-		return array(
-			'objects' => $results->orders,
-			'total'   => $results->total,
-			'pages'   => $results->max_num_pages,
-		);
 	}
 
 	/**
